Move terminal entry class definition to seperate file
